small fixes

+++ server/src/utils/file.utils.js
@@ -3,10 +3,6 @@ const path = require("path");
 
 class FileUtils {
   uploadFile(file, fileName) {
-    const validation = fileValidation(file);
-    if (!validation) {
-      return false;
-    }
     file.mv(path.resolve(__dirname, "../static/", fileName), (err) => {
       if (err) throw err;
       console.log("File created successfully");
@@ -14,20 +10,13 @@ class FileUtils {
   }
 
   deleteFile(path) {
-    fs.unlink(path, (err) => {
-      if (err) throw err;
-      console.log("File deleted successfully");
-    });
+    if (fs.existsSync(path)) {
+      fs.unlink(path, (err) => {
+        if (err) throw err;
+        console.log("File deleted successfully");
+      });
+    }
   }
 }
-
-const fileValidation = (file) => {
-  const allowingExtensions = /(\.jpg|\.jpeg|\.png|\.svg|\.bmp)&/i;
-  if (!allowingExtensions.exec(file)) {
-    console.log("Invalid file type");
-    return false;
-  }
-  return true;
-};
 
 module.exports = new FileUtils();
